feat/add-get-withdraw

+++ internal/domain/gophermart/entities/entities.go
@@ -18,6 +18,6 @@ type Order struct {
 
 type Withdraw struct {
 	OrderNumber string    `json:"order" db:"order_number"`
-	Sum         int       `json:"sum" db:"sum"`
+	Sum         float64   `json:"sum" db:"sum"`
 	UploadedAt  time.Time `json:"processed_at" db:"upload_time"`
 }
